summarizer_googlecloud/pyproject.toml
[project]
name = "summarizer-googlecloud"
version = "0.1.0"
description = "Google Cloud Gemini API based summarizer"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "tenacity",
    "pyyaml",
]

[project.scripts]
summarizer-googlecloud = "summarizer_googlecloud:main"
